Refactor turbomind (low-level abstractions) (#3423)
* low level abstraction

* refactor

* eliminate template

* remove unused

* refactor bindings

* simplify lm head

* refactor weight

* fix tp

* cublas

* refactor sampling

* remove unused

* simplify

* fix AWQ support

* fix moe

* fix nccl lm_head

* fix

* refactor data types

* skip legacy ut

* simplify

* rename data types

* refactor

* refactor runtime states

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* format

* remove unused

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix msvc build

* fix ut & msvc build

* fix ut & msvc build

* fix gcc build

* fix lint & ut

* fix lint

* fetch Catch2 when building tests

* rewind msvc build

* fix sampling

+++ src/turbomind/kernels/sampling_topk_kernels.h
@@ -48,10 +48,10 @@ void invokeCurandInitialize(curandState_t*     state,
                             unsigned long long random_seed,
                             cudaStream_t       stream);
 
-void invokeCurandBatchInitialize(curandState_t*            states,
-                                 const size_t              batch_size,
-                                 const unsigned long long* random_seeds,
-                                 cudaStream_t              stream);
+void invokeCurandBatchInitialize(curandState_t*  states,
+                                 const size_t    batch_size,
+                                 const uint64_t* random_seeds,
+                                 cudaStream_t    stream);
 
 struct TopKSortFilterParams {
     void*  workspace;
